Resolve Whitaker Dylint findings across the crate and its tests
Prepare the codebase for adopting the Whitaker Dylint suite by fixing
every finding it reports, without suppressions:

- bumpy_road_function: extract build_agent_overrides and
  serialize_agent_override from build_overrides in the config loader,
  and append_file_entry/append_symlink_entry from
  append_non_directory_entry in the credential archive writer.
- module_max_lines: split the protocol exec module into stdin_io and
  output_io sibling submodules; move the error module's tests to a
  sibling error_tests.rs; split the github tests into client_tests and
  property_tests submodules; split the ACP frame and protocol ACP test
  modules; move the configuration layer-precedence BDD steps into
  tests/bdd_config_steps/layer.rs.
- module_must_have_inner_docs: add //! inner documentation to every
  unit-test module that lacked it.
- no_expect_outside_tests: helper and fixture functions no longer
  panic; they return Result (io::Result, serde_json::Error, or
  PodbotError) and propagate failures, with the calling test bodies
  unwrapping as the test verdict. Mutex-guard recorders in test
  doubles now recover from poisoning via PoisonError::into_inner
  instead of panicking.
- no_std_fs_operations: the Makefile audit-target test and the BDD
  config-loader helpers now perform fixture filesystem work through
  cap-std directory handles instead of ambient std::fs calls.

Also resolve the Clippy too_many_arguments and needless_pass_by_value
diagnostics introduced by the refactoring.
