+++ src/engine/connection/exec/protocol.rs
@@ -60,18 +60,14 @@
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use std::time::Duration;
 
 use bollard::container::LogOutput;
 use bollard::errors::Error as BollardError;
-use futures_util::{Stream, StreamExt};
-use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
-use tokio::task::JoinHandle;
-use tokio::time::timeout;
+use futures_util::Stream;
+use tokio::io::{AsyncRead, AsyncWrite};
 
 use super::ExecRequest;
 use super::acp_frame::OutboundFrameAssembler;
-use super::acp_helpers;
 use super::acp_policy::MethodDenylist;
 use super::acp_runtime::{
     OutboundPolicyAdapter, SINK_CHANNEL_CAPACITY, WriteCmd, run_container_stdin_sink,
@@ -81,6 +77,16 @@ use super::host_io::stdin_forwarding_disabled_for_tests;
 use super::runtime_helpers::exec_failed;
 use super::session::CapabilityPolicy;
 use crate::error::PodbotError;
+
+#[path = "protocol_output.rs"]
+mod output_io;
+#[path = "protocol_stdin.rs"]
+mod stdin_io;
+
+use self::output_io::{AdapterOutputIo, run_output_loop_async, run_output_loop_with_adapter};
+use self::stdin_io::{
+    forward_host_stdin_to_channel, forward_host_stdin_to_exec_async, settle_stdin_forwarding_task,
+};
 /// Host-side stdio handles used by the protocol byte proxy.
 pub(super) struct ProtocolProxyIo<HostStdin, HostStdout, HostStderr> {
     /// Host stdin reader supplied to the forwarding task.
@@ -92,14 +98,6 @@ pub(super) struct ProtocolProxyIo<HostStdin, HostStdout, HostStderr> {
     /// Per-session configuration knobs applied to this proxy run.
     options: ProtocolSessionOptions,
 }
-
-/// Allow a short grace period for EOF- and flush-driven completion paths to
-/// finish before treating stdin forwarding as stalled. `50ms` matches typical
-/// local pipe and socket flush timings in this code path without adding a
-/// noticeable shutdown delay; if future benchmarks or transport changes show
-/// different behaviour, adjust `STDIN_SETTLE_TIMEOUT` and extend the proxy
-/// tests that exercise EOF and non-EOF stdin shutdown cases.
-const STDIN_SETTLE_TIMEOUT: Duration = Duration::from_millis(50);
 
 /// Maximum bytes buffered between host stdin reads and container input writes.
 /// This bounds memory consumption per read cycle and provides backpressure by
@@ -339,259 +337,6 @@ where
         })
 }
 
-/// Reads the first newline-delimited ACP frame from `buffered_stdin`,
-/// applies capability masking, and forwards the resulting bytes to
-/// `sender` as a [`WriteCmd::Forward`].
-///
-/// Returns `Ok(true)` when the caller should continue pumping host
-/// stdin (either because the masker produced no bytes to forward, or
-/// because the send to the sink succeeded). Returns `Ok(false)` when
-/// the sink channel has closed and the caller should return cleanly.
-async fn send_masked_initialize_frame<R>(
-    buffered_stdin: &mut tokio::io::BufReader<R>,
-    sender: &tokio::sync::mpsc::Sender<WriteCmd>,
-) -> io::Result<bool>
-where
-    R: AsyncRead + Unpin,
-{
-    let bytes = acp_helpers::read_and_mask_initial_acp_frame(buffered_stdin).await?;
-    if bytes.is_empty() {
-        return Ok(true);
-    }
-    Ok(sender.send(WriteCmd::Forward(bytes)).await.is_ok())
-}
-
-/// Pumps the remainder of host stdin into the container-stdin sink as
-/// a sequence of [`WriteCmd::Forward`] chunks bounded by
-/// [`STDIN_BUFFER_CAPACITY`]. Stops on EOF or when the sink channel
-/// closes.
-async fn pump_raw_frames<R>(
-    buffered_stdin: &mut tokio::io::BufReader<R>,
-    sender: &tokio::sync::mpsc::Sender<WriteCmd>,
-) -> io::Result<()>
-where
-    R: AsyncRead + Unpin,
-{
-    use tokio::io::AsyncReadExt;
-
-    let mut buf = vec![0u8; STDIN_BUFFER_CAPACITY];
-    loop {
-        let bytes_read = buffered_stdin.read(&mut buf).await?;
-        if bytes_read == 0 {
-            break;
-        }
-        let chunk = buf
-            .get(..bytes_read)
-            .map(<[u8]>::to_vec)
-            .unwrap_or_default();
-        if sender.send(WriteCmd::Forward(chunk)).await.is_err() {
-            break;
-        }
-    }
-    Ok(())
-}
-async fn forward_host_stdin_to_channel<HostStdin>(
-    host_stdin: HostStdin,
-    sender: tokio::sync::mpsc::Sender<WriteCmd>,
-    rewrite_acp_initialize: bool,
-) -> io::Result<()>
-where
-    HostStdin: AsyncRead + Unpin,
-{
-    let mut buffered_stdin = tokio::io::BufReader::with_capacity(STDIN_BUFFER_CAPACITY, host_stdin);
-
-    if rewrite_acp_initialize && !send_masked_initialize_frame(&mut buffered_stdin, &sender).await?
-    {
-        return Ok(());
-    }
-
-    pump_raw_frames(&mut buffered_stdin, &sender).await
-}
-
-struct AdapterOutputIo<'a, HostStdout, HostStderr> {
-    adapter: &'a mut OutboundPolicyAdapter,
-    host_stdout: &'a mut HostStdout,
-    host_stderr: &'a mut HostStderr,
-}
-async fn run_output_loop_with_adapter<HostStdout, HostStderr>(
-    container_id: &str,
-    output: &mut Pin<Box<dyn Stream<Item = Result<LogOutput, BollardError>> + Send>>,
-    io: &mut AdapterOutputIo<'_, HostStdout, HostStderr>,
-) -> Result<(), PodbotError>
-where
-    HostStdout: AsyncWrite + Unpin,
-    HostStderr: AsyncWrite + Unpin,
-{
-    while let Some(chunk_result) = output.next().await {
-        let chunk = chunk_result
-            .map_err(|error| exec_failed(container_id, format!("exec stream failed: {error}")))?;
-        match chunk {
-            LogOutput::StdOut { message } | LogOutput::Console { message } => {
-                io.adapter
-                    .handle_chunk(message.as_ref(), io.host_stdout)
-                    .await
-                    .map_err(|error| {
-                        exec_failed(
-                            container_id,
-                            format!("failed writing stdout output: {error}"),
-                        )
-                    })?;
-            }
-            LogOutput::StdErr { message } => {
-                write_output_chunk(container_id, io.host_stderr, message.as_ref(), "stderr")
-                    .await?;
-            }
-            LogOutput::StdIn { .. } => {}
-        }
-    }
-    Ok(())
-}
-/// Wait for the stdin forwarding task to complete within a short grace
-/// period.
-async fn settle_stdin_forwarding_task(
-    container_id: &str,
-    mut stdin_task: JoinHandle<io::Result<()>>,
-    options: ProtocolSessionOptions,
-) -> Result<(), PodbotError> {
-    let Ok(join_result) = timeout(STDIN_SETTLE_TIMEOUT, &mut stdin_task).await else {
-        // The container output path has already completed, so stdin can be
-        // cancelled instead of waiting indefinitely on a live host reader. A
-        // timeout here still indicates that stdin forwarding did not complete
-        // cleanly before shutdown, so protocol mode must surface that failure
-        // instead of reporting success with potentially truncated input.
-        abort_stdin_forwarding_task(stdin_task);
-        if options.disable_stdin_forwarding || stdin_forwarding_disabled_for_tests() {
-            return Ok(());
-        }
-        return Err(exec_failed(
-            container_id,
-            "stdin forwarding did not complete before protocol session shutdown",
-        ));
-    };
-
-    classify_stdin_forwarding_task_result(container_id, join_result)
-}
-
-/// Map a join result from the stdin forwarding task to a `PodbotError`.
-fn classify_stdin_forwarding_task_result(
-    container_id: &str,
-    join_result: Result<io::Result<()>, tokio::task::JoinError>,
-) -> Result<(), PodbotError> {
-    match join_result {
-        Ok(Ok(())) => Ok(()),
-        Ok(Err(error)) => Err(exec_failed(
-            container_id,
-            format!("failed forwarding stdin to exec input: {error}"),
-        )),
-        Err(error) if error.is_cancelled() => Ok(()),
-        Err(error) => Err(exec_failed(
-            container_id,
-            format!("stdin forwarding task failed: {error}"),
-        )),
-    }
-}
-
-/// Abort and drop the stdin forwarding task without awaiting it.
-fn abort_stdin_forwarding_task(stdin_task: JoinHandle<io::Result<()>>) {
-    if !stdin_task.is_finished() {
-        stdin_task.abort();
-        // Avoid awaiting the aborted task here because host stdin may be
-        // blocked in a non-cancellable read. Dropping the handle mirrors the
-        // attached-session shutdown path and keeps teardown bounded.
-        drop(stdin_task);
-    }
-}
-
-/// Copy host stdin to the container exec input, optionally rewriting the
-/// first ACP `initialize` frame before the raw copy begins.
-async fn forward_host_stdin_to_exec_async<HostStdin>(
-    host_stdin: HostStdin,
-    mut input: Pin<Box<dyn AsyncWrite + Send>>,
-    rewrite_acp_initialize: bool,
-) -> io::Result<()>
-where
-    HostStdin: AsyncRead + Unpin,
-{
-    let mut buffered_stdin = tokio::io::BufReader::with_capacity(STDIN_BUFFER_CAPACITY, host_stdin);
-
-    if rewrite_acp_initialize {
-        acp_helpers::forward_initial_acp_frame_async(&mut buffered_stdin, &mut input).await?;
-    }
-    tokio::io::copy(&mut buffered_stdin, &mut input).await?;
-
-    input.flush().await?;
-    input.shutdown().await
-}
-
 #[cfg(test)]
 #[path = "protocol_acp_tests.rs"]
 mod acp_tests;
-
-/// Drain the container output stream, routing each chunk to host stdout or
-/// stderr.
-async fn run_output_loop_async<HostStdout, HostStderr>(
-    container_id: &str,
-    output: &mut Pin<Box<dyn Stream<Item = Result<LogOutput, BollardError>> + Send>>,
-    host_stdout: &mut HostStdout,
-    host_stderr: &mut HostStderr,
-) -> Result<(), PodbotError>
-where
-    HostStdout: AsyncWrite + Unpin,
-    HostStderr: AsyncWrite + Unpin,
-{
-    while let Some(chunk_result) = output.next().await {
-        let chunk = chunk_result
-            .map_err(|error| exec_failed(container_id, format!("exec stream failed: {error}")))?;
-        handle_log_output_chunk(container_id, chunk, host_stdout, host_stderr).await?;
-    }
-
-    Ok(())
-}
-
-/// Route a single container log-output chunk to the appropriate host stream.
-async fn handle_log_output_chunk<HostStdout, HostStderr>(
-    container_id: &str,
-    chunk: LogOutput,
-    host_stdout: &mut HostStdout,
-    host_stderr: &mut HostStderr,
-) -> Result<(), PodbotError>
-where
-    HostStdout: AsyncWrite + Unpin,
-    HostStderr: AsyncWrite + Unpin,
-{
-    match chunk {
-        LogOutput::StdOut { message } | LogOutput::Console { message } => {
-            write_output_chunk(container_id, host_stdout, message.as_ref(), "stdout").await
-        }
-        LogOutput::StdErr { message } => {
-            write_output_chunk(container_id, host_stderr, message.as_ref(), "stderr").await
-        }
-        LogOutput::StdIn { .. } => Ok(()),
-    }
-}
-
-/// Write and flush a byte slice to `writer`, mapping I/O failures to
-/// `PodbotError`.
-async fn write_output_chunk<Writer>(
-    container_id: &str,
-    writer: &mut Writer,
-    bytes: &[u8],
-    stream_name: &str,
-) -> Result<(), PodbotError>
-where
-    Writer: AsyncWrite + Unpin,
-{
-    writer.write_all(bytes).await.map_err(|error| {
-        exec_failed(
-            container_id,
-            format!("failed writing {stream_name} output: {error}"),
-        )
-    })?;
-    writer.flush().await.map_err(|error| {
-        exec_failed(
-            container_id,
-            format!("failed flushing {stream_name} output: {error}"),
-        )
-    })?;
-    Ok(())
-}
