+++ src/engine/connection/exec/protocol_acp_policy_integration_tests.rs
@@ -2,7 +2,7 @@
 
 use std::io;
 use std::pin::Pin;
-use std::sync::{Arc, Mutex};
+use std::sync::{Arc, Mutex, PoisonError};
 use std::task::{Context, Poll};
 
 use bollard::container::LogOutput;
@@ -21,7 +21,10 @@ struct RecordingOutput {
 
 impl RecordingOutput {
     fn snapshot(&self) -> Vec<u8> {
-        self.bytes.lock().expect("recording mutex").clone()
+        self.bytes
+            .lock()
+            .unwrap_or_else(PoisonError::into_inner)
+            .clone()
     }
 }
 
@@ -33,7 +36,7 @@ impl AsyncWrite for RecordingOutput {
     ) -> Poll<io::Result<usize>> {
         self.bytes
             .lock()
-            .expect("recording mutex")
+            .unwrap_or_else(PoisonError::into_inner)
             .extend_from_slice(buf);
         Poll::Ready(Ok(buf.len()))
     }
@@ -55,55 +58,57 @@ fn protocol_request() -> Result<ExecRequest, PodbotError> {
     )
 }
 
-fn blocked_request_frame(id: i64) -> Vec<u8> {
+fn blocked_request_frame(id: i64) -> Result<Vec<u8>, serde_json::Error> {
     let mut bytes = serde_json::to_vec(&serde_json::json!({
         "jsonrpc": "2.0",
         "id": id,
         "method": "terminal/create",
         "params": {},
-    }))
-    .expect("blocked request should serialize");
+    }))?;
     bytes.push(b'\n');
-    bytes
+    Ok(bytes)
 }
 
-fn drive_policy_session(policy: CapabilityPolicy) -> (Vec<u8>, Vec<u8>, Vec<u8>) {
-    let runtime = tokio::runtime::Runtime::new().expect("runtime should build");
-    let initialize = super::initialize_frame("\n");
-    let host_stdin = runtime
-        .block_on(super::build_host_stdin(&initialize))
-        .expect("host stdin should build");
+fn drive_policy_session(policy: CapabilityPolicy) -> io::Result<(Vec<u8>, Vec<u8>, Vec<u8>)> {
+    let runtime = tokio::runtime::Runtime::new()?;
+    let initialize = super::initialize_frame("\n").map_err(io::Error::other)?;
+    let host_stdin = runtime.block_on(super::build_host_stdin(&initialize))?;
     let host_stdout = RecordingOutput::default();
     let host_stdout_handle = host_stdout.clone();
     let host_stderr = RecordingOutput::default();
     let host_stderr_handle = host_stderr.clone();
     let container_input = super::RecordingInputWriter::new();
     let container_stdin = container_input.bytes.clone();
+    let blocked = blocked_request_frame(7).map_err(io::Error::other)?;
     let output = stream::iter([Ok(LogOutput::StdOut {
-        message: blocked_request_frame(7).into(),
+        message: blocked.into(),
     })]);
+    let request = protocol_request().map_err(io::Error::other)?;
     let stdio = ProtocolProxyIo::new(host_stdin, host_stdout, host_stderr)
         .with_options(ProtocolSessionOptions::new().with_capability_policy(policy));
 
     runtime
         .block_on(run_protocol_session_with_io_async(
-            &protocol_request().expect("protocol request should build"),
+            &request,
             Box::pin(output),
             Box::pin(container_input),
             stdio,
         ))
-        .expect("policy session should complete");
+        .map_err(io::Error::other)?;
 
-    (
-        container_stdin.lock().expect("stdin mutex").clone(),
+    let recorded_container_stdin = container_stdin
+        .lock()
+        .unwrap_or_else(PoisonError::into_inner)
+        .clone();
+    Ok((
+        recorded_container_stdin,
         host_stdout_handle.snapshot(),
         host_stderr_handle.snapshot(),
-    )
+    ))
 }
 
-fn parse_json_line(bytes: &[u8]) -> serde_json::Value {
+fn parse_json_line(bytes: &[u8]) -> Result<serde_json::Value, serde_json::Error> {
     serde_json::from_slice(bytes.strip_suffix(b"\n").unwrap_or(bytes))
-        .expect("line should contain JSON")
 }
 
 fn split_lines(bytes: &[u8]) -> Vec<&[u8]> {
@@ -117,8 +122,12 @@ fn line_matches(bytes: &[u8], expected: &[u8]) -> bool {
     bytes == expected
 }
 
+/// Pure query: returns `true` when `bytes` parses as the synthesized denial
+/// response for the blocked request with id 7 and the given `method`.
 fn synthesized_response_for_method(bytes: &[u8], method: &str) -> bool {
-    let response = parse_json_line(bytes);
+    let Ok(response) = parse_json_line(bytes) else {
+        return false;
+    };
     response.get("id") == Some(&serde_json::json!(7))
         && response
             .get("error")
@@ -130,9 +139,10 @@ fn synthesized_response_for_method(bytes: &[u8], method: &str) -> bool {
 #[test]
 fn mask_and_deny_masks_initialize_and_synthesizes_blocked_response() {
     let (container_stdin, host_stdout, _host_stderr) =
-        drive_policy_session(CapabilityPolicy::MaskAndDeny);
+        drive_policy_session(CapabilityPolicy::MaskAndDeny).expect("policy session should run");
     let lines = split_lines(&container_stdin);
-    let expected_initialize = super::mask_acp_initialize_frame(&super::initialize_frame("\n"));
+    let initialize = super::initialize_frame("\n").expect("initialize frame should serialize");
+    let expected_initialize = super::mask_acp_initialize_frame(&initialize);
 
     assert!(
         host_stdout.is_empty(),
@@ -155,12 +165,12 @@ fn mask_and_deny_masks_initialize_and_synthesizes_blocked_response() {
 #[test]
 fn disabled_policy_preserves_plain_streaming_proxy_behaviour() {
     let (container_stdin, host_stdout, _host_stderr) =
-        drive_policy_session(CapabilityPolicy::Disabled);
-    let blocked = blocked_request_frame(7);
+        drive_policy_session(CapabilityPolicy::Disabled).expect("policy session should run");
+    let initialize = super::initialize_frame("\n").expect("initialize frame should serialize");
+    let blocked = blocked_request_frame(7).expect("blocked request should serialize");
 
     assert_eq!(
-        container_stdin,
-        super::initialize_frame("\n"),
+        container_stdin, initialize,
         "Disabled should forward host stdin without ACP masking",
     );
     assert_eq!(
@@ -172,12 +182,13 @@ fn disabled_policy_preserves_plain_streaming_proxy_behaviour() {
 #[test]
 fn mask_only_masks_initialize_without_runtime_denylist_enforcement() {
     let (container_stdin, host_stdout, _host_stderr) =
-        drive_policy_session(CapabilityPolicy::MaskOnly);
-    let blocked = blocked_request_frame(7);
+        drive_policy_session(CapabilityPolicy::MaskOnly).expect("policy session should run");
+    let initialize = super::initialize_frame("\n").expect("initialize frame should serialize");
+    let blocked = blocked_request_frame(7).expect("blocked request should serialize");
 
     assert_eq!(
         container_stdin,
-        super::mask_acp_initialize_frame(&super::initialize_frame("\n")),
+        super::mask_acp_initialize_frame(&initialize),
         "MaskOnly should apply the initialize masking path",
     );
     assert_eq!(
